Split the File module into smaller parts
